+++ fleetpings/constants.py
@@ -2,4 +2,7 @@
 Constants
 """
 
+APP_NAME = "aa-fleetpings"
+GITHUB_URL = f"https://github.com/ppfeufer/{APP_NAME}"
+
 DISCORD_WEBHOOK_REGEX = r"https:\/\/discord\.com\/api\/webhooks\/[\d]+\/[a-zA-Z0-9_-]+$"